+++ mycodo/inputs/system_cpuload.py
@@ -1,6 +1,8 @@
 # coding=utf-8
 import os
 
+from flask_babel import lazy_gettext
+
 from mycodo.inputs.base_input import AbstractInput
 
 # Measurements